Integrate redundant methods into one class
